add snack to _app.js
